refactor(ReimbursementStatusAction): Split into multiple files
This make it more easy to maintain, test and override
